move github stuff to its own package
