+++ src/github/model.rs
@@ -1,0 +1,190 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Root {
+    #[serde(rename = "ref")]
+    pub push_ref: String,
+
+    // The commit hash before this push
+    pub before: String,
+
+    // The commit hash after this push
+    pub after: String,
+
+    pub repository: Repository,
+
+    pub sender: Sender,
+
+    pub head_commit: HeadCommit,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Repository {
+    pub id: u64,
+
+    pub name: String,
+    pub full_name: String,
+    pub owner: Owner,
+    pub description: Option<String>,
+    pub fork: bool,
+    pub created_at: u64,
+    pub updated_at: String,
+    pub pushed_at: u64,
+    pub default_branch: String,
+    pub master_branch: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Owner {
+    pub id: u64,
+
+    pub name: String,
+    pub email: String,
+    pub login: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Sender {
+    pub id: u64,
+
+    pub login: String,
+    pub url: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Commit {
+    pub id: String,
+
+    pub tree_id: String,
+    pub distinct: bool,
+    pub message: String,
+    pub timestamp: String,
+
+    // pub author: Author,
+    pub committer: Committer,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct CommitAuthor {
+    pub name: String,
+    pub email: String,
+    pub username: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Committer {
+    pub name: String,
+    pub email: String,
+    pub username: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct HeadCommit {
+    pub id: String,
+    pub tree_id: String,
+    pub distinct: bool,
+    pub message: String,
+    pub timestamp: String,
+    pub url: String,
+    pub author: CommitAuthor,
+    pub committer: Committer,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct GetReleaseRoot {
+    pub id: u64,
+    pub url: String,
+    pub assets_url: String,
+    pub upload_url: String,
+    pub tarball_url: String,
+    pub zipball_url: String,
+    pub tag_name: String,
+    pub name: String,
+    pub body: String,
+    pub draft: bool,
+    pub prerelease: bool,
+    pub created_at: String,
+    pub published_at: String,
+    pub author: Author,
+    pub assets: Vec<ReleaseAsset>,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Author {
+    pub login: String,
+    pub id: u64,
+    pub node_id: String,
+    pub avatar_url: String,
+    pub gravatar_id: String,
+    pub url: String,
+    pub html_url: String,
+    pub followers_url: String,
+    pub following_url: String,
+    pub gists_url: String,
+    pub starred_url: String,
+    pub subscriptions_url: String,
+    pub organizations_url: String,
+    pub repos_url: String,
+    pub events_url: String,
+    pub received_events_url: String,
+    #[serde(rename = "type")]
+    pub type_field: String,
+    pub site_admin: bool,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct ReleaseAsset {
+    pub url: String,
+    pub browser_download_url: String,
+    pub id: u64,
+    pub node_id: String,
+    pub name: String,
+    pub label: String,
+    pub state: String,
+    pub content_type: String,
+    pub size: u64,
+    pub download_count: u64,
+    pub created_at: String,
+    pub updated_at: String,
+    pub uploader: Uploader,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Uploader {
+    pub login: String,
+    pub id: u64,
+    pub node_id: String,
+    pub avatar_url: String,
+    pub gravatar_id: String,
+    pub url: String,
+    pub html_url: String,
+    pub followers_url: String,
+    pub following_url: String,
+    pub gists_url: String,
+    pub starred_url: String,
+    pub subscriptions_url: String,
+    pub organizations_url: String,
+    pub repos_url: String,
+    pub events_url: String,
+    pub received_events_url: String,
+    #[serde(rename = "type")]
+    pub type_field: String,
+    pub site_admin: bool,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct UploadReleaseAssetRoot {
+    pub url: String,
+    pub browser_download_url: String,
+    pub id: u64,
+    pub node_id: String,
+    pub name: String,
+    pub label: String,
+    pub state: String,
+    pub content_type: String,
+    pub size: u64,
+    pub download_count: u64,
+    pub created_at: String,
+    pub updated_at: String,
+    pub uploader: Uploader,
+}
